+++ src/main/java/domain/Side.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.coordinate.Direction;
+
 public enum Side {
 
     HAN,
