+++ src/main/java/domain/board/BoardInitializer.java
@@ -1,6 +1,6 @@
 package domain.board;
 
-import domain.Position;
+import domain.coordinate.Position;
 import domain.Side;
 import domain.piece.Piece;
 
@@ -9,5 +9,6 @@ import java.util.Map;
 public interface BoardInitializer {
 
     Map<Position, Piece> initialize();
+
     Side getFirstTurnSide();
 }
