+++ src/main/java/domain/coordinate/Direction.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.coordinate;
 
 public enum Direction {
