+++ src/test/java/domain/piece/HorseTest.java
@@ -3,7 +3,7 @@ package domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.Game;
-import domain.Position;
+import domain.coordinate.Position;
 import domain.Side;
 import domain.board.BoardInitializer;
 import java.util.HashMap;
